Fixing warnings
Mostly transitioning unused variables to TODO comments.
